adicionei preco com virgula
arrumei o script do fooditem e draganddrop e adicionei numeros com virgula, arrumei os text nas fases e adicionei a fase 5

+++ Assets/Scripts/FoodItem.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class FoodItem : MonoBehaviour
 {
-    public int value; // Valor do alimento
+    public float value; // Valor do alimento (agora com casas decimais)
 
-    public void SetFoodValue(int newValue)
+    public void SetFoodValue(float newValue)
     {
         value = newValue;
     }
